refactor: Create NodeIdProvider interface
With an interface, we can create a "static" implementation to be used
when the nodeId is set in the configuration
